% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquachem_variables.R
\name{aquachem_variables}
\alias{aquachem_variables}
\title{Retrieve unique fysicochemical variables from aquachem database.}
\usage{
aquachem_variables()
}
\description{
The function can only be used from within the INBO network.
}
