% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_aquachem.R
\name{connect_aquachem}
\alias{connect_aquachem}
\title{Connect to the INBO aquachem database}
\usage{
connect_aquachem()
}
\description{
Returns a connection to the INBO aquachem database.
The function can only be used from within the INBO network.
}
