% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_vmm.R
\name{connect_vmm}
\alias{connect_vmm}
\title{Connect to the local copy of the VMM database}
\usage{
connect_vmm()
}
\description{
Returns a connection to the INBO aquachem database.

The function can only be used from within the INBO network.
}
