% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chem_aquachem.R
\name{get_chem_aquachem}
\alias{get_chem_aquachem}
\title{Retrieve fysicochemical data from water samples in the aquachem database.}
\usage{
get_chem_aquachem(
  locs,
  con = connect(aquachem),
  variable = NULL,
  collect = FALSE
)
}
\arguments{
\item{locs}{exported locations from get_locs_aquachem command}

\item{con}{connection to aquachem database}

\item{variable}{search specific variable(s). To get an overview of the existing variables use aquachem_variables().}

\item{collect}{set to TRUE to collect data}
}
\description{
The function can only be used from within the INBO network.
}
\examples{
\dontrun{get_chem_aquachem(locs,con, variable = c("NH4","PO4","NO3"))}
}
