% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chem_vmm.R
\name{get_chem_vmm}
\alias{get_chem_vmm}
\title{retrieve fysicochemical data from the local VMM-database
The function can only be used from within the INBO network.}
\usage{
get_chem_vmm(
  con,
  bbox = NULL,
  variable = NULL,
  stream = NULL,
  guess = FALSE,
  mask = NULL,
  buffer_mask = NULL,
  stream_geodatabase = FALSE,
  buffer_stream = NULL,
  geodatabase = NULL,
  full_output = FALSE
)
}
\arguments{
\item{con}{connection to local VMM database}

\item{bbox}{search within bbox}

\item{variable}{define variable(s)}

\item{stream}{define river name}

\item{guess}{set to TRUE will search for stream string}

\item{mask}{search within shapefile (.shp)}

\item{buffer_mask}{create additional mask around buffer}

\item{stream_geodatabase}{collect data from stream segment geodatabase. Geodatabase needs to contain vhag_code (vhag_code) AND/OR river name (NAAM) and spatial data of river segment.}

\item{buffer_stream}{create search buffer around river segments (not all VMM water sample points intersect with stream segments)}

\item{geodatabase}{link to the geodatabase}

\item{full_output}{give complete output}
}
\description{
retrieve fysicochemical data from the local VMM-database
The function can only be used from within the INBO network.
}
