% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locs_aquachem.R
\name{get_locs_aquachem}
\alias{get_locs_aquachem}
\title{Retrieve locations from the aquachem database}
\usage{
get_locs_aquachem(
  con = connect_aquachem(),
  mask = NULL,
  buffer = 0,
  bbox = NULL,
  province = NULL,
  town = NULL,
  habtype_f = NULL,
  collect = FALSE
)
}
\arguments{
\item{con}{connection to the aquachem database, default is connect_aquachem().}

\item{mask}{search within shapefile (.shp)}

\item{buffer}{create and include additional buffer around mask (in m)}

\item{bbox}{search within bbox, format = c("xmax", "xmin", "ymax", "ymin")}

\item{province}{search within province}

\item{town}{search within town}

\item{habtype_f}{search habitattype}

\item{collect}{collect data from lazy query}
}
\description{
The function can only be used from within the INBO network.
}
